Move `FeedViewAdapter` to it's own file
